Revert "#86a3fu0eu - ref: Refactor repository that updates change date"
