+++ tests/unit/infra/repositories/user/UpdateTimestampRepository.spec.ts
@@ -1,13 +1,13 @@
 import sinon from 'sinon';
 import HttpRequestErrors from 'src/domains/common/helpers/HttpRequestErrors';
 import { HttpStatusCode } from 'src/domains/common/helpers/HttpStatusCode';
-import UpdateTimestampRepository from 'src/infra/repositories/common/UpdateTimestampRepository';
+import UpdateTimestampRepository from 'src/infra/repositories/user/UpdateTimestampRepository';
 import { Logger } from 'src/types/shared/logger';
 
 describe('Infra :: Repositories :: User :: UpdateTimestampRepository', () => {
     let updateTimestampRepository: UpdateTimestampRepository, database: any;
 
-    const logger: Logger = () => { };
+    const logger: Logger = () => {};
 
     context('#updateTimestamp', () => {
         const findOne = sinon.spy(() => ({ userId: '123', email: 'test@email.com' }));
